Initial implementations of ks_get_kek().  Untested, and none of these are secure (the one in ks_flash.c is a stub, and the others are for cases where we have no secure hardware in which to store the KEK).
These are primarily for testing, since in the long run the entire
software implementation of AES-keywrap will be replaced by Verilog
which never lets software see the unwrapped key.  Or so says current
theory.  For the moment, we just need something that will let us test
the rest of the RPC and keystore mechanisms.