CRC-32 code for use in ks_flash, stm32 DFU, possibly elsewhere.
This is an open source C99 CRC-32 implementation generated by pycrc,
see notes in source on copyright status and pycrc options used.

crc32.c contains two different implementations of the CRC-32 algorithm
with the same API, one optimized for speed, the other optimized for
much smaller code space at the expense of speed.  We use the fast
implementation by default, but maybe the small implementation will be
useful, eg, in the bootloader.  Remove the extra later if this turns
out to have been a waste of time.