First cut at multi-attribute get/set/delete API.
This is not yet complete, only the ks_volatile driver supports it,
ks_flash will be a bit more complicated and isn't written yet.

At the moment, this adds a complete duplicate set of
{set,get,delete}_attributes() functions in parallel to the earlier
{set,get,delete}_attribute() functions.  We will almost certainly want
to get rid of the duplicates, probably (but not necessarily) the
entire single-attribute suite.  At the moment, though, we want both
sets so we can compare execution speeds of the two sets of functions.