Clean up libhal build system a bit.
Exporting CFLAGS as an environment variable turns out to interact
badly with certain other settings here.  I *think* this only happens
when we use one of the shorthand targets which re-runs make in the
same directory with non-default settings, but this is complicated
enough without having to remember which voodoo triggers it.  So
instead of exporting CFLAGS as an environment variable we just pass it
on the command line in the handful of cases where it's needed.