Backport ECDSA core support to ksng branch.
Support for the core/pkey/ecdsa{256,384} cores is cooked before the
branch on which it was based.  Oops.  Time to backport.

See pymux branch for original commit history.  git should do the right
thing when the pymux branch is cooked enough to merge back to the ksng
or master branches.