Work around known bugs in PyCrypto ASN.1 code.
Turns out there are a couple of known minor bugs in PyCrypto's ASN.1
decoder, simple dumb things that never could have worked.  Debian's
packaging includes a patch for these bugs, but for some reason the
patch is marked as not needing to be sent upstream, dunno why.  So
these methods work fine, but only on Debian.  Feh.

Simplest approach is to work around the bugs on all platforms,
particularly given that this is just unit test support code.