First pass on experimental one-size-fits-nobody keystore.
Support for variable-length keystore objects significantly complicates
the keystore implementation, including serious some serious code bloat
and a complex recovery algorithm to deal with crashes or loss of power
at exactly the wrong time.  Perhaps we don't really need this?

So this is an experiment to see whether we can replace variable-length
keystore objects with fixed-length, perhaps with a compile time option
to let us make the fixed object length be 8192 bytes instead of 4096
bytes when needed to hold things like large RSA keys.

First pass on this is just throwing away nearly 1,000 lines of
excessively complex code.  The result probably won't even compile yet,
but it's already significantly easier to read.