Automatic logout when client disconnects or muxd restarts.
The HSM itself should be detecting carrier drop on its RPC port, but I
haven't figured out where the DCD bit is hiding in the STM32 UART API,
and the muxd has to be involved in this in any case, since only the
muxd knows when an individual client connection has dropped.  So, for
the moment, we handle all of this in the muxd.