First cut at reusing RSA blinding factors.
General technique here suggested by Peter Gutman.  If I got the math
wrong, blame me, not Peter.

This has not yet been tested to confirm that it returns correct
results when using the blinding factors cache, and preliminary timing
results suggest that we may be chasing the wrong performance problem.
Unclear whether we'll ever really want to integrate this change, but
pushing it on a branch to get it into repository history.

If we do end up using this, the blinding factors cache will need minor
redesign, principally to use the external SDRAM because main memory
has gotten kind of full.  Some way to clear the cache when restarting
the HSM would be nice, probably requires a `hal_rsa_init()` function.