Release keystore lock before unwrapping key.
hal_ks_fetch() was written as lock-at-the-top, unlock-at-the-bottom to
keep it as simple as possible, but this turns out to have bad
performance implications when unwrapping the key is slow.  So now we
grab the wrapped key, release the lock, then unwrap, which should be
safe enough given that hal_ks_fetch() is read-only.  This lets us make
better use of multiple AES cores to unwrap in parallel when we have
multiple active clients.