Add small cache for RSA blinding factors.
Generating new RSA blinding factors turns out to be relatively
expensive, but we can amortize that cost by maintaining a small cache
and simply mutating old values after each use with a cheaper
operation.  Squaring works, pretty much by definition.

Blinding factors are only sort-of-sensitive: we don't want them to
leak out of the HSM, but they're only based on the public modulus, not
the private key components, and we're only using them to foil side
channel attacks, so the risk involved in caching them seems small.

For the moment, the cache is very small, since we only care about this
for bulk signature operations.  Tune this later if it becomes an issue.