Disable seldom-used FMC I/O debugging code by default.
Profiling reports significant time spent in the hal_io_fmc.c debugging
code even when runtime debugging is off.  This is odd, and may be a
profiling artifact, but we don't use that debugging code often, so if
it costs anything at all we might as well disable it when not needed.